+++ node/src/openshift-dotnet/Uhuru.Openshift.Utilities/RubyHash.cs
@@ -7,6 +7,30 @@ namespace Uhuru.Openshift.Utilities
 {
     public class RubyHash : Dictionary<string, object>
     {
+        public RubyHash()
+            : base()
+        {
+
+        }
+
+        public RubyHash(Dictionary<string, object> content)
+            : this()
+        {
+            foreach (var record in content)
+            {
+                this[record.Key] = record.Value;
+            }
+        }
+
+        public RubyHash(Dictionary<string, string> content)
+            : this()
+        {
+            foreach (var record in content)
+            {
+                this[record.Key] = record.Value;
+            }
+        }
+
         new public dynamic this[string key]
         {
             get
@@ -51,6 +75,23 @@ namespace Uhuru.Openshift.Utilities
         }
 
         public RubyHash Merge(Dictionary<string, object> another)
+        {
+            RubyHash result = new RubyHash();
+
+            foreach (string key in this.Keys)
+            {
+                result[key] = this[key];
+            }
+
+            foreach (string key in another.Keys)
+            {
+                result[key] = another[key];
+            }
+
+            return result;
+        }
+
+        public RubyHash Merge(Dictionary<string, string> another)
         {
             RubyHash result = new RubyHash();